To debug offline
git-svn-id: http://svn.cern.ch/guest/lhcb/Online/trunk@99663 4525493e-7705-40b1-a816-d608a930855b

Former-commit-id: 02db7fa623a4388ed4173a3fa592af849442c4cc

+++ Online/Online/ExampleGaucho/src/GaudiExample.h
@@ -71,8 +71,8 @@ private:
   std::vector<double> m_vectDouble;
   
   // counters into rates test
-  double count_rate_1;
-  double count_rate_2;
+  int count_rate_1;
+  int count_rate_2;
     
   int nbinEntries;
   double sumOfWeights;
